Use vanilla https module instead of request; ES6
Yay! Request was nice in the beginning to get this kickstarted, as I
was already familiar with its API, though I find it to be quite a
'heavy' module, especially for this use case where it's very
straightforward what needs to be done, so I am happy to have swapped it
out for Node's plain https module. In fact, it's nearly just as much
'work' to implement in the `request` module than with Node's https.request.